feat: handle session expiration for opened protected note for both plain advanced editors
